Replicate https://stackoverflow.com/questions/77705053
This commit introduces a new module "sof-questions-77705053" to the existing application. The module is set up with its own pom.xml and application properties, and also included several new classes such as DocumentEntity, DocumentRepository, DocumentController, and DocumentEntityRequest. The application main pom.xml file is also updated to include this module.